+++ CHANGELOG.md
@@ -95,3 +95,4 @@
     * show SocketError on screens without internet
     * remove extraneous files and code
     * move showToast method to Screen mixin
+    * move floatingActionButton to Screen mixin